services/orientation/requirements.txt
onnxruntime>=1.16.0
Pillow>=10.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
